pyproject.toml
# MIT License
#
# Copyright (C) 2024 The Little Hackers.  All rights reserved.
#
# Permission is hereby granted, free of charge, to any person obtaining
# a copy of this software and associated documentation files (the
# "Software"), to deal in the Software without restriction, including
# without limitation the rights to use, copy, modify, merge, publish,
# distribute, sublicense, and/or sell copies of the Software, and to
# permit persons to whom the Software is furnished to do so, subject to
# the following conditions:
#
# The above copyright notice and this permission notice shall be
# included in all copies or substantial portions of the Software.
#
# THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND,
# EXPRESS OR IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF
# MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT.
# IN NO EVENT SHALL THE AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY
# CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION OF CONTRACT,
# TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION WITH THE
# SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.

[tool.poetry]
authors = ["Daniel CAUNE <daniel.caune@thelittlehackers.com>"]
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "License :: Other/Proprietary License",
    "Operating System :: OS Independent",
    "Programming Language :: Python",
    "Topic :: Software Development :: Libraries",
]
description = "Python library helper for reading environment variables"
include = [
    "CHANGELOG.md"
]
keywords = ["env", "library", "thelittlehackers"]
license = "SEE LICENSE IN <LICENSE.md>"
name = "perseus-getenv-library"
packages = [{ include = "thelittlehackers", from = "src" }]
readme = "README.md"
repository = "https://github.com/thelittlehackers-devops/library_getenv_python"
version = "1.0.4"

[tool.poetry.dependencies]
python = "^3.10"
python-dotenv = "^1.0.0"
thelittlehackers-core = {git = "git@github.com:the-little-hackers/library_thelittlehackers-core_python.git"}

[tool.poetry.group.dev.dependencies]
black = "^24.8.0"
flake8 = "^7.1.1"
setuptools = "^74.0.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
